requirements.txt
aiohappyeyeballs==2.4.3
aiohttp==3.11.6
aiohttp-retry==2.8.3
aiosignal==1.3.1
asgiref==3.8.1
attrs==24.2.0
certifi==2024.8.30
charset-normalizer==3.4.0
Django==5.1.3
django-environ==0.11.2
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
propcache==0.2.0
PyJWT==2.10.0
requests==2.32.3
sqlparse==0.5.1
twilio==9.3.7
urllib3==2.2.3
whitenoise==6.8.2
yarl==1.17.2

openai
psycopg2-binary